+++ encore_api_cli/commands/draw.py
@@ -1,8 +1,10 @@
+import json
 from pathlib import Path
 from urllib.parse import urlparse
 
 import click
 
+from encore_api_cli.exceptions import ClickException
 from encore_api_cli.options import common_options
 from encore_api_cli.output import write_message, write_success
 from encore_api_cli.state import pass_state
@@ -11,6 +13,8 @@ from encore_api_cli.utils import color_id, color_path, get_client
 
 def draw_options(f):
     """Set draw options."""
+    f = click.option("--no-download", is_flag=True, help="Disable download.")(f)
+    f = click.option("--rule", help="Drawing rules.")(f)
     f = click.option(
         "-o",
         "--out_dir",
@@ -19,7 +23,6 @@ def draw_options(f):
         show_default=True,
         help="Path of directory to output drawn file.",
     )(f)
-    f = click.option("--no-download", is_flag=True, help="Disable download")(f)
     return f
 
 
@@ -33,10 +36,10 @@ def cli() -> None:  # noqa: D103
 @draw_options
 @common_options
 @pass_state
-def draw(state, keypoint_id, out_dir, no_download):
+def draw(state, keypoint_id, out_dir, rule, no_download):
     """Draw keypoints on uploaded movie or image."""
     c = get_client(state)
-    drawing_id = c.draw_keypoint(keypoint_id)
+    drawing_id = c.draw_keypoint(keypoint_id, rule=_parse_rule(rule))
     write_message(f"Drawing started. (drawing_id: {color_id(drawing_id)})")
 
     # TODO: invoke download command
@@ -61,3 +64,18 @@ def draw(state, keypoint_id, out_dir, no_download):
         write_message("Drawing is timed out.")
     else:
         write_message("Drawing failed.")
+
+
+def _parse_rule(rule):
+    if rule is None:
+        return rule
+
+    try:
+        rule = json.loads(rule)
+    except json.decoder.JSONDecodeError:
+        raise ClickException("Rule format is invalid. Must be in JSON format.")
+
+    if not isinstance(rule, list):
+        raise ClickException("Rule format is invalid. Must be in list format.")
+
+    return rule